+++ extended/src/test/java/apoc/ml/OpenAITestResultUtils.java
@@ -10,14 +10,13 @@ public class OpenAITestResultUtils {
     public static final String EMBEDDING_QUERY = "CALL apoc.ml.openai.embedding(['Some Text'], $apiKey, $conf)";
     public static final String CHAT_COMPLETION_QUERY = """
             CALL apoc.ml.openai.chat([
-            {role:"system", content:"Only answer with a single word"},
+            {role:"system", content:"You are a dictionary that answer with a single word"},
             {role:"user", content:"What planet do humans live on?"}
             ], $apiKey, $conf)
             """;
     public static final String CHAT_COMPLETION_QUERY_WITHOUT_SYSTEM = """
             CALL apoc.ml.openai.chat([
-            { content: "What planet do humans live on?", role: "user" },
-            { content: "Only answer with a single word", role: "assistant" }
+            { content: "What planet do humans live on? Only answer with a single word", role: "user" }
             ], $apiKey, $conf)
             """;
     public static final String COMPLETION_QUERY = "CALL apoc.ml.openai.completion('What color is the sky? Answer in one word: ', $apiKey, $conf)";
