+++ extra-dependencies/build.gradle
@@ -77,7 +77,7 @@ ext {
 
 
 subprojects {
-    version = '2025.04.0'
+    version = '2025.05.0'
     group = 'org.neo4j.contrib'
 }
